# api-spec.yaml
openapi: 3.0.3
info:
  title: Go Search Engine API
  description: |
    A high-performance, full-text search engine API with advanced features including:
    - Full-text search with typo tolerance
    - Prefix search and autocomplete
    - Advanced filtering and ranking
    - Unicode support
    - Document indexing and management
  version: 1.0.0
  contact:
    name: Go Search Engine
    url: https://github.com/gcbaptista/go-search-engine
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /indexes:
    post:
      tags:
        - Index Management
      summary: Create a new search index
      description: Creates a new search index with specified configuration including searchable fields, filterable fields, and ranking criteria.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndexSettings"
            examples:
              basic_index:
                summary: Basic index for movies
                value:
                  name: "movies"
                  searchable_fields: ["title", "cast", "genres"]
                  filterable_fields: ["year", "rating", "genres"]
                  ranking_criteria:
                    - field: "popularity"
                      order: "desc"
                    - field: "rating"
                      order: "desc"
                  min_word_size_for_1_typo: 4
                  min_word_size_for_2_typos: 7
              advanced_index:
                summary: Advanced index with prefix search disabled
                value:
                  name: "documents"
                  searchable_fields: ["title", "content", "author"]
                  filterable_fields: ["category", "date_created", "status"]
                  ranking_criteria:
                    - field: "date_created"
                      order: "desc"
                  min_word_size_for_1_typo: 3
                  min_word_size_for_2_typos: 6
                  fields_without_prefix_search: ["author"]
                  no_typo_tolerance_fields: ["author"]
      responses:
        "201":
          description: Index created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
              example:
                message: "Index 'movies' created successfully"
        "400":
          description: Invalid request body or missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missing_name:
                  summary: Missing index name
                  value:
                    error: "Index name is required"
                invalid_json:
                  summary: Invalid JSON format
                  value:
                    error: "Invalid request body: invalid character '}' looking for beginning of value"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Failed to create index: disk write error"

    get:
      tags:
        - Index Management
      summary: List all indexes
      description: Retrieves a list of all available search indexes.
      responses:
        "200":
          description: List of indexes retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  indexes:
                    type: array
                    items:
                      type: string
                    description: Array of index names
                  count:
                    type: integer
                    description: Total number of indexes
              example:
                indexes: ["movies", "documents", "products"]
                count: 3
  /indexes/{indexName}:
    get:
      tags:
        - Index Management
      summary: Get index details
      description: Retrieves detailed information about a specific index including its configuration settings.
      parameters:
        - name: indexName
          in: path
          required: true
          description: Name of the index to retrieve
          schema:
            type: string
          example: "movies"
      responses:
        "200":
          description: Index details retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndexSettings"
              example:
                name: "movies"
                searchable_fields: ["title", "cast", "genres"]
                filterable_fields: ["year", "rating", "genres"]
                ranking_criteria:
                  - field: "popularity"
                    order: "desc"
                min_word_size_for_1_typo: 4
                min_word_size_for_2_typos: 7
                fields_without_prefix_search: []
        "404":
          description: Index not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Index 'movies' not found"

    delete:
      tags:
        - Index Management
      summary: Delete an index
      description: Permanently deletes an index and all its associated data.
      parameters:
        - name: indexName
          in: path
          required: true
          description: Name of the index to delete
          schema:
            type: string
          example: "movies"
      responses:
        "200":
          description: Index deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
              example:
                message: "Index 'movies' deleted successfully"
        "404":
          description: Index not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Index 'movies' not found"
        "500":
          description: Failed to delete index
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Failed to delete index 'movies': permission denied"
  /indexes/{indexName}/stats:
    get:
      tags:
        - Index Management
      summary: Get index statistics
      description: Retrieves detailed statistics about a specific index including document count and configuration
      parameters:
        - name: indexName
          in: path
          required: true
          description: Name of the index to get statistics for
          schema:
            type: string
          example: "movies"
      responses:
        "200":
          description: Index statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    description: Index name
                  document_count:
                    type: integer
                    description: Total number of documents in the index
                  searchable_fields:
                    type: array
                    items:
                      type: string
                    description: Fields that can be searched
                  filterable_fields:
                    type: array
                    items:
                      type: string
                    description: Fields that can be used in filters
                  typo_settings:
                    type: object
                    properties:
                      min_word_size_for_1_typo:
                        type: integer
                      min_word_size_for_2_typos:
                        type: integer
                  field_settings:
                    type: object
                    properties:
                      fields_without_prefix_search:
                        type: array
                        items:
                          type: string
                      no_typo_tolerance_fields:
                        type: array
                        items:
                          type: string
                      distinct_field:
                        type: string
              example:
                name: "movies"
                document_count: 1250
                searchable_fields: ["title", "cast", "genres"]
                filterable_fields: ["year", "rating", "genres"]
                typo_settings:
                  min_word_size_for_1_typo: 4
                  min_word_size_for_2_typos: 7
                field_settings:
                  fields_without_prefix_search: []
                  no_typo_tolerance_fields: ["genres"]
                  distinct_field: "title"
        "404":
          description: Index not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Index 'movies' not found"
  /indexes/{indexName}/settings:
    patch:
      tags:
        - Index Management
      summary: Update index settings
      description: Updates specific settings of an existing index. Currently supports updating fields_without_prefix_search.
      parameters:
        - name: indexName
          in: path
          required: true
          description: Name of the index to update
          schema:
            type: string
          example: "movies"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndexSettingsUpdate"
            example:
              fields_without_prefix_search: ["author", "isbn"]
              no_typo_tolerance_fields: ["isbn", "product_code"]
      responses:
        "200":
          description: Settings updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  warning:
                    type: string
              example:
                message: "Settings updated successfully for index 'movies'"
                warning: "You may need to reindex your documents for changes to FieldsWithoutPrefixSearch to take full effect."
        "400":
          description: Invalid request or no changes detected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Index not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to update settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    IndexSettings:
      type: object
      required:
        - name
        - searchable_fields
      properties:
        name:
          type: string
          description: Unique name for the index
          example: "movies"
        searchable_fields:
          type: array
          items:
            type: string
          description: |
            Fields that can be searched, in priority order. 
            IMPORTANT: Order matters! The search engine will fully exhaust each field 
            (exact matches + typo tolerance) before proceeding to the next field.
          example: ["title", "cast", "genres"]
        filterable_fields:
          type: array
          items:
            type: string
          description: Fields that can be used in filters
          example: ["year", "rating", "genres"]
        ranking_criteria:
          type: array
          items:
            $ref: "#/components/schemas/RankingCriterion"
          description: Ordered list of ranking criteria
        min_word_size_for_1_typo:
          type: integer
          description: Minimum word length to allow 1 typo
          default: 4
          example: 4
        min_word_size_for_2_typos:
          type: integer
          description: Minimum word length to allow 2 typos
          default: 7
          example: 7
        fields_without_prefix_search:
          type: array
          items:
            type: string
          description: Fields for which prefix/n-gram search is disabled
          example: ["author", "isbn"]
        no_typo_tolerance_fields:
          type: array
          items:
            type: string
          description: Fields for which typo tolerance is disabled (only exact matches)
          example: ["isbn", "product_code"]
        distinct_field:
          type: string
          description: Field to use for deduplication to avoid returning duplicate documents
          example: "title"
    RankingCriterion:
      type: object
      required:
        - field
        - order
      properties:
        field:
          type: string
          description: Field name to rank by
          example: "popularity"
        order:
          type: string
          enum: ["asc", "desc"]
          description: Sort order
          example: "desc"
    IndexSettingsUpdate:
      type: object
      properties:
        fields_without_prefix_search:
          type: array
          items:
            type: string
          description: Fields for which prefix/n-gram search should be disabled
          example: ["author", "isbn"]
        no_typo_tolerance_fields:
          type: array
          items:
            type: string
          description: Fields for which typo tolerance should be disabled (only exact matches)
          example: ["isbn", "product_code"]
        distinct_field:
          type: string
          description: Field to use for deduplication to avoid returning duplicate documents
          example: "title"
...
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
          description: Success message
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message describing what went wrong
          example: "Invalid request body"
...